Move bigtable/internal/row_key_sample.h to bigtable/row_key_sample.h
