+++ google/cloud/bigtable/bigtable_client.bzl
@@ -42,7 +42,6 @@ bigtable_client_HDRS = [
     "internal/poll_longrunning_operation.h",
     "internal/prefix_range_end.h",
     "internal/readrowsparser.h",
-    "internal/row_key_sample.h",
     "internal/rpc_policy_parameters.inc",
     "internal/rpc_policy_parameters.h",
     "internal/rowreaderiterator.h",
@@ -55,6 +54,7 @@ bigtable_client_HDRS = [
     "polling_policy.h",
     "read_modify_write_rule.h",
     "row.h",
+    "row_key_sample.h",
     "row_range.h",
     "row_reader.h",
     "row_set.h",
